+++ hydromt_wflow/wflow.py
@@ -2499,7 +2499,7 @@ either {'temp' [°C], 'temp_min' [°C], 'temp_max' [°C], 'wind' [m/s], 'rh' [%]
             Reprojection method from rasterio.enums.Resampling. to reproject the climate
             data to the model resolution. By default 'nearest_index'.
         fillna_method: str, optional
-            Method to fill NaN cells within the active model domain in the 
+            Method to fill NaN cells within the active model domain in the
             temperature data e.g. 'nearest'
             By default None for no interpolation.
         chunksize: int, optional
@@ -3158,37 +3158,100 @@ Run setup_soilmaps first"
             self.set_config(option, states_config[option])
 
     # I/O
-    def read(self):
+    def read(
+        self,
+    ):
         """Read the complete model schematization and configuration from file."""
         self.read_config()
         self.read_grid()
+        self.read_geoms()
+        self.read_forcing()
         self.read_intbl()
         self.read_tables()
-        self.read_geoms()
         self.read_states()
-        self.read_forcing()
         self.logger.info("Model read")
 
-    def write(self):
-        """Write the complete model schematization and configuration to file."""
+    def write(
+        self,
+        config_fn: str = None,
+        grid_fn: Union[Path, str] = "staticmaps.nc",
+        geoms_fn: Union[Path, str] = "staticgeoms",
+        forcing_fn: Union[Path, str] = None,
+        states_fn: Union[Path, str] = None,
+    ):
+        """
+        Write the complete model schematization and configuration to file.
+
+        From this function, the output filenames/folder of the different components can
+        be set. If not set, the default filenames/folder are used.
+        To change more advanced settings, use the specific write methods directly.
+
+        Parameters
+        ----------
+        config_fn : str, optional
+            Name of the config file, realtive to model root. By default None.
+        grid_fn : str, optional
+            Name of the grid file, relative to model root/dir_input. By default
+            'staticmaps.nc'.
+        geoms_fn : str, optional
+            Name of the geoms folder relative to grid_fn (ie model root/dir_input). By
+            default 'staticgeoms'.
+        forcing_fn : str, optional
+            Name of the forcing file relative to model root/dir_input. By default None
+            to use the name as defined in the model config file.
+        states_fn : str, optional
+            Name of the states file realtive to model root/dir_input. By default None
+            to use the name as defined in the model config file.
+        """
         self.logger.info(f"Write model data to {self.root}")
         # if in r, r+ mode, only write updated components
         if not self._write:
             self.logger.warning("Cannot write in read-only mode")
             return
         self.write_data_catalog()
-        if self.config:  # try to read default if not yet set
-            self.write_config()
+        _ = self.config  # try to read default if not yet set
         if self._grid:
-            self.write_grid()
+            self.write_grid(fn_out=grid_fn)
+        if self._geoms:
+            self.write_geoms(geoms_fn=geoms_fn)
+        if self._forcing:
+            self.write_forcing(fn_out=forcing_fn)
         if self._tables:
             self.write_tables()
-        if self._geoms:
-            self.write_geoms()
         if self._states:
-            self.write_states()
-        if self._forcing:
-            self.write_forcing()
+            self.write_states(states_fn=states_fn)
+        # Write the config last as variables can get set in other write methods
+        self.write_config(config_name=config_fn)
+
+    def write_config(
+        self,
+        config_name: Optional[str] = None,
+        config_root: Optional[str] = None,
+    ):
+        """
+        Write config to <root/config_fn>.
+
+        Parameters
+        ----------
+        config_name : str, optional
+            Name of the config file. By default None to use the default name
+            wflow_sbm.toml.
+        config_root : str, optional
+            Root folder to write the config file if different from model root (default).
+        """
+        self._assert_write_mode()
+        if config_name is not None:
+            self._config_fn = config_name
+        elif self._config_fn is None:
+            self._config_fn = self._CONF
+        if config_root is None:
+            config_root = self.root
+        fn = join(config_root, self._config_fn)
+        # Create the folder if it does not exist
+        if not isdir(dirname(fn)):
+            os.makedirs(dirname(fn))
+        self.logger.info(f"Writing model config to {fn}")
+        self._configwrite(fn)
 
     def read_grid(self, **kwargs):
         """
@@ -3247,7 +3310,9 @@ Run setup_soilmaps first"
         Parameters
         ----------
         fn_out : Path | str, optional
-            Name or path to the outgoing grid file (including extension).
+            Name or path to the outgoing grid file (including extension). This is the
+            path/name relative to the root folder and if present the ``dir_input``
+            folder.
         """
         if not self._write:
             raise IOError("Model opened in read-only mode")
@@ -3303,22 +3368,29 @@ Run setup_soilmaps first"
 
     def read_geoms(
         self,
-        geom_fn: str = "staticgeoms",
+        geoms_fn: str = "staticgeoms",
     ):
         """
         Read static geometries and adds to ``geoms``.
 
-        Assumes that the `geom_fn` folder is located in the same folder as the static
-        input data (``input.path_static``). If not found uses assumes they are in
-        <root/geom_fm>.
+        Assumes that the `geoms_fn` folder is located relative to root and ``dir_input``
+        if defined in the toml. If not found, uses assumes they are in <root/geoms_fn>.
+
+        Parameters
+        ----------
+        geoms_fn : str, optional
+            Folder name/path where the static geometries are stored relative to the
+            model root and ``dir_input`` if any. By default "staticgeoms".
         """
         if not self._write:
             self._geoms = dict()  # fresh start in read-only mode
-        dir_default = join(self.root, "staticmaps.nc")
-        dir_mod = dirname(
-            self.get_config("input.path_static", abs_path=True, fallback=dir_default)
-        )
-        fns = glob.glob(join(dir_mod, geom_fn, "*.geojson"))
+        # Check if dir_input is set and add
+        if self.get_config("dir_input") is not None:
+            dir_mod = join(self.get_config("dir_input", abs_path=True), geoms_fn)
+        else:
+            dir_mod = join(self.root, geoms_fn)
+
+        fns = glob.glob(join(dir_mod, "*.geojson"))
         if len(fns) > 1:
             self.logger.info("Reading model staticgeom files.")
         for fn in fns:
@@ -3328,10 +3400,22 @@ Run setup_soilmaps first"
 
     def write_geoms(
         self,
-        geom_fn: str = "staticgeoms",
+        geoms_fn: str = "staticgeoms",
         precision: int | None = None,
     ):
-        """Write geoms in <root/geom_fn> in GeoJSON format."""
+        """
+        Write geoms in GeoJSON format.
+
+        Checks the path of ``geoms_fn`` using both model root and
+        ``dir_input``. If not found uses the default path ``staticgeoms`` in the root
+        folder.
+
+        Parameters
+        ----------
+        geoms_fn : str, optional
+            Folder name/path where the static geometries are stored relative to the
+            model root and ``dir_input`` if any. By default "staticgeoms".
+        """
         # to write use self.geoms[var].to_file()
         if not self._write:
             raise IOError("Model opened in read-only mode")
@@ -3345,6 +3429,15 @@ Run setup_soilmaps first"
                 else:
                     _precision = 6
             grid_size = 10 ** (-_precision)
+            # Prepare the output folder
+            if self.get_config("dir_input") is not None:
+                geoms_dir = join(
+                    self.get_config("dir_input", abs_path=True),
+                    geoms_fn,
+                )
+            else:
+                geoms_dir = join(self.root, geoms_fn)
+
             for name, gdf in self.geoms.items():
                 # TODO change to geopandas functionality once geopandas 1.0.0 comes
                 # See https://github.com/geopandas/geopandas/releases/tag/v1.0.0-alpha1
@@ -3352,7 +3445,7 @@ Run setup_soilmaps first"
                     gdf.geometry,
                     grid_size=grid_size,
                 )
-                fn_out = join(self.root, geom_fn, f"{name}.geojson")
+                fn_out = join(geoms_dir, f"{name}.geojson")
                 gdf.to_file(fn_out, driver="GeoJSON")
 
     def read_forcing(self):
@@ -3451,7 +3544,6 @@ see https://pandas.pydata.org/pandas-docs/stable/user_guide/timeseries.html#offs
             # get output filename
             if fn_out is not None:
                 self.set_config("input.path_forcing", fn_out)
-                self.write_config()  # re-write config
             else:
                 fn_name = self.get_config("input.path_forcing", abs_path=False)
                 if fn_name is not None:
@@ -3516,7 +3608,6 @@ change name input.path_forcing "
                         return
                     else:
                         self.set_config("input.path_forcing", fn_default)
-                        self.write_config()  # re-write config
                         fn_out = fn_default_path
 
             # Check if all dates between (starttime, endtime) are in all da forcing
@@ -3544,9 +3635,9 @@ change name input.path_forcing "
                     f"Not all dates found in precip_fn changing starttime to \
 {start} and endtime to {end} in the toml."
                 )
-                self.set_config("starttime", start.to_pydatetime())
-                self.set_config("endtime", end.to_pydatetime())
-                self.write_config()
+                # Set the strings first
+                self.set_config("starttime", start.strftime("%Y-%m-%dT%H:%M:%S"))
+                self.set_config("endtime", end.strftime("%Y-%m-%dT%H:%M:%S"))
 
             if decimals is not None:
                 ds = ds.round(decimals)
@@ -3588,7 +3679,6 @@ change name input.path_forcing "
                 fns_out = os.path.relpath(fn_out, self.root)
                 fns_out = f"{str(fns_out)[0:-3]}_*.nc"
                 self.set_config("input.path_forcing", fns_out)
-                self.write_config()  # re-write config
                 for label, ds_gr in ds.resample(time=freq_out):
                     # ds_gr = group[1]
                     start = ds_gr["time"].dt.strftime("%Y%m%d")[0].item()